pyproject.toml
[project]
name = "misalud-entendida"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.7.0",
    "gradio>=6.4.0",
    "huggingface-hub>=0.36.0",
    "modal>=1.3.0.post1",
    "pillow>=12.1.0",
    "pypdf>=6.1.3",
    "requests>=2.28.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
exclude = ["medgemma_official_example.py"]
